Change length type to float

+++ my-ACG/create-items/create-anime-item.py
@@ -46,7 +46,7 @@ def main():
     parser.add_argument('--seen', type=int, default=0)
     parser.add_argument('--episodes', type=int, default=0)
     parser.add_argument('--status', type=int, choices=[0, 1, 2], default=0)
-    parser.add_argument('--length', type=int, default=24)
+    parser.add_argument('--length', type=float, default=24)
     parser.add_argument('--wp')
     parser.add_argument('--moe')
     parser.add_argument('--gamer')
